One functional Test Case for a C# BatchLoader
First cut of a C# BatchLoader, with the initial functional test case passing.

Much of the BatchLoader is left untouched, such as loading of Issues, which gets a bit deeper into parsing the XML.

To this end, we needed to talk to the DB, so these changes directly add Dapper as our ORM, along with some plumbing code.