WIP Adding skeleton dotnet CLI project
Right now this accepts the arguments needed for running batch load, but it doesn't do anything with them.

We still need to call the actual loader, but before that we need to figure out how we're resolving our dependencies.

So, not functional in its current form, but compiling and running.